Refactor RailsBase into a module
